Refactoring for formats directory
