+++ lib/cloudevent.js
@@ -1,6 +1,6 @@
 var Spec_0_1          = require('./specs/spec_0_1.js');
 var Spec_0_2          = require('./specs/spec_0_2.js');
-var JSONFormatter_0_1 = require('./format/json_0_1.js');
+var JSONFormatter_0_1 = require('./formats/json_0_1.js');
 
 function Cloudevent(_spec, _formatter){
   this.spec      = (_spec) ? new _spec(Cloudevent) : new Spec_0_1(Cloudevent);
